services/typing-pdf-extractor-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
minio==7.2.5
pymupdf==1.23.26
pydantic==2.6.3
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv
